// src/app/interfaces/Workers.ts
export interface Workers {
  id: number;
  idCardNumber: string;
  name: string;
  lastName: string;
  mobile: string;
  email: string;
  jobTypeId: number;
  picture: string;
  deleted: boolean;
  empNumber: string;
  address: string;
  cityId: number;
  workerStatusId: number;
  statusReport: string;
}